Fix oom_instead_of_bump_pointer_overflow breakage
Nightly changes tp the `Layout` API have caused this test to break as
layouts are now restricted to a max size of `isize::MAX`. This is fixed
by using `isize::MAX` instead of `usize::MAX` in the test.

+++ tests/tests.rs
@@ -71,7 +71,7 @@ fn oom_instead_of_bump_pointer_overflow() {
     let p = x as *mut u8 as usize;
 
     // A size guaranteed to overflow the bump pointer.
-    let size = usize::MAX - p + 1;
+    let size = (isize::MAX as usize) - p + 1;
     let align = 1;
     let layout = match Layout::from_size_align(size, align) {
         Err(e) => {
